feat(models): add StrictStructuredOutput base; relocate provider-safe mixin
New low-level module reflexio/models/structured_output.py holds
_fold_oneof_to_anyof, ProviderSafeUnionMixin, and the StrictStructuredOutput
base (hook only, no shared config) so both models/ and server/ layers can
inherit it without a layering inversion. PlaybookConsolidationOutput now
inherits StrictStructuredOutput; llm_utils keeps its own structure-aware inline
fold in make_strict_json_schema (deliberately not the blind helper).
